able to add albums. also did some css. not looking great but not painful. also made header and footer templates
