--- src/main/java/com/chaseatucker/songr/GreetingController.java
@@ -1,12 +1,0 @@
-package com.chaseatucker.songr;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class GreetingController {
-  @GetMapping("/hello")
-  public String hello() {
-    return "hello";
-  }
-}
