+++ src/main/java/com/chaseatucker/songr/model/AlbumRepository.java
@@ -1,0 +1,7 @@
+package com.chaseatucker.songr.model;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AlbumRepository extends JpaRepository<Album, Long> {
+
+}
