+++ src/main/resources/application.properties
@@ -1,1 +1,5 @@
-
+spring.datasource.url=jdbc:postgresql://localhost:5432/songr
+spring.datasource.username=tuckerc
+spring.datasource.password=ozgbb6CX7mD6d8hQhHH3
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto = update
