+++ src/main/resources/static/css/layout.css
@@ -1,8 +1,6 @@
 body {
   display: grid;
-  grid-template: 15% 80% 5% / 1fr;
-  width: 100vw;
-  height: 100vh;
+  grid-template: auto auto auto / 1fr;
   justify-items: stretch;
   align-items: stretch;
 }
@@ -15,28 +13,74 @@ header {
 }
 
 main {
-  margin-top: 3vh;
   display: grid;
-  grid-template: 20% 80% / 100%;
+  grid-template: auto auto auto auto / 1fr;
 }
 
 main h2 {
+  text-align: center;
+  vertical-align: middle;
+  grid-row: 3 / 4;
+  grid-column: auto;
+}
+
+main h3 {
   text-align: center;
   vertical-align: middle;
   grid-row: 1 / 2;
   grid-column: auto;
 }
 
+main form {
+  display: grid;
+  grid-row: 2 / 3;
+  grid-column: auto;
+  justify-self: stretch;
+  align-self: stretch;
+  grid-template: auto / auto auto auto auto auto auto;
+  justify-items: stretch;
+  align-items: stretch;
+}
+
+form input[name="title"] {
+  grid-row: auto;
+  grid-column: 1 / 2;
+}
+
+form input[name="artist"] {
+  grid-row: auto;
+  grid-column: 2 / 3;
+}
+
+form input[name="songCount"] {
+  grid-row: auto;
+  grid-column: 3 / 4;
+}
+
+form input[name="length"] {
+  grid-row: auto;
+  grid-column: 4 / 5;
+}
+
+form input[name="imageUrl"] {
+  grid-row: auto;
+  grid-column: 5 / 6;
+}
+
+form input[name="submit"] {
+  grid-row: auto;
+  grid-column: 6 / 7;
+}
+
 main ul {
   text-align: center;
-  grid-row: 2 / 3;
+  grid-row: 4 / 5;
   grid-column: auto;
   justify-self: stretch;
   align-self: stretch;
 }
 
 main li {
-  display: inline-block;
   text-align: center;
   padding: 0.5vh 2vw;
 }
