+++ src/main/resources/templates/albums.html
@@ -8,24 +8,28 @@
     <link rel="stylesheet" type="text/css" href="css/theme.css">
 </head>
 <body>
-    <header>
-        <h1>Songr</h1>
-        <p>The definitive source for song stuff!</p>
-    </header>
+    <header th:replace="fragments/fragments :: header"></header>
     <main>
-        <h2>These Three Albums</h2>
+        <h3>Add an Album to the Collection</h3>
+        <h2>All The Albums</h2>
+        <form action="/albums" method="POST">
+            <label>Title:<input type="text" name="title" required></label>
+            <label>Artist:<input type="text" name="artist" required></label>
+            <label>Song Count:<input type="number" name="songCount" required></label>
+            <label>Length:<input type="number" name="length" required></label>
+            <label>Image URL:<input type="text" name="imageUrl" required></label>
+            <input type="submit" name="submit">
+        </form>
         <ul>
             <li th:each="album : ${albums}">
-                <h3 th:text="'Album Title: ' + ${album.title}"></h3>
-                <h4 th:text="'Artist ' + ${album.artist}"></h4>
+                <h4 th:text="'Album Title: ' + ${album.title}"></h4>
+                <h5 th:text="'Artist ' + ${album.artist}"></h5>
                 <p th:text="'Number of Songs: ' + ${album.songCount}"></p>
                 <p th:text="'Album Length: ' + ${album.length} + ' seconds'"></p>
                 <img th:alt="${album.title}" th:src="${album.imageUrl}">
             </li>
         </ul>
     </main>
-    <footer>
-        <p>&copy; 2020 songr.com<p></p>
-    </footer>
+    <footer th:replace="fragments/fragments :: footer"></footer>
 </body>
 </html>
