+++ src/main/resources/templates/fragments/fragments.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Fragments</title>
+</head>
+<body>
+  <header th:fragment="header">
+    <h1>Songr</h1>
+    <p>The definitive source for song stuff!</p>
+  </header>
+
+  <footer th:fragment="footer">
+    <p>&copy; 2020 songr.com<p></p>
+  </footer>
+
+</body>
+</html>
